windowing/gbm: split legacy drm
